challenge2-solution/image_api/requirements.txt
fastapi
uvicorn
pydantic
numpy
pandas
opencv-python
Pillow
